moved flies to maintain order in the project
