+++ src/__test__/hooks/useDirection.test.tsx
@@ -1,8 +1,8 @@
 import { cleanup, fireEvent, render } from "@testing-library/react";
 import { afterEach, beforeAll, describe, expect, it } from "vitest";
-import ScrollWatcher from "../lib/components/ScrollWatcher";
-import UseDirection from "./components/direction/UseDirection";
-import matchMediaMock from "./mock/matchMediaMock";
+import ScrollWatcher from "../../lib/components/ScrollWatcher";
+import UseDirection from "../test-components/direction/UseDirection";
+import matchMediaMock from "../mock/matchMediaMock";
 
 describe("useDirection hook", () => {
   afterEach(cleanup);
